% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desjardins7_e2sfca.R
\name{DJ7_WC_BASE}
\alias{DJ7_WC_BASE}
\title{Primary cumulative Gaussian band weights (30/60/120/180 min).}
\usage{
DJ7_WC_BASE
}
\description{
Primary cumulative Gaussian band weights (30/60/120/180 min).
}
