% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_step_floating_catchment.R
\name{.sci_join_msg}
\alias{.sci_join_msg}
\title{Standard message for an unmatched scientific join key.}
\usage{
.sci_join_msg(lhs, rhs, by, ids, extra = "")
}
\description{
Standard message for an unmatched scientific join key.
}
\keyword{internal}
