% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_step_floating_catchment.R
\name{.unmatched_keys}
\alias{.unmatched_keys}
\title{Keys present on the left of a join with no match on the right.}
\usage{
.unmatched_keys(x, y, by)
}
\description{
Keys present on the left of a join with no match on the right.
}
\keyword{internal}
