% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_step_floating_catchment.R
\name{E2SFCA_DEFAULT_THRESHOLDS}
\alias{E2SFCA_DEFAULT_THRESHOLDS}
\title{Default access thresholds (in per-100k units) for cell-level population shares.}
\usage{
E2SFCA_DEFAULT_THRESHOLDS
}
\description{
Default access thresholds (in per-100k units) for cell-level population shares.
}
