% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_bands.R
\name{get_canonical_bands}
\alias{get_canonical_bands}
\title{Return the canonical generation bands (minutes).}
\usage{
get_canonical_bands()
}
\value{
[CANONICAL_BANDS].
}
\description{
Return the canonical generation bands (minutes).
}
