% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_bands.R
\name{get_primary_access_band}
\alias{get_primary_access_band}
\title{Return the primary access band in the requested units.}
\usage{
get_primary_access_band(units = c("min", "sec"))
}
\arguments{
\item{units}{`"min"` (default) or `"sec"`.}
}
\value{
Integer scalar -- [PRIMARY_ACCESS_BAND_MIN] or [PRIMARY_ACCESS_BAND_SEC].
}
\description{
Return the primary access band in the requested units.
}
