% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urps_inpatient_sigma.R
\name{URPS_INPATIENT_SIGMA_BY_AVAILABILITY_MIN}
\alias{URPS_INPATIENT_SIGMA_BY_AVAILABILITY_MIN}
\title{The same stratified bandwidths expressed in minutes}
\format{
Named numeric vector, sigma in minutes.
}
\usage{
URPS_INPATIENT_SIGMA_BY_AVAILABILITY_MIN
}
\description{
The same stratified bandwidths expressed in minutes
}
